+++ gba_test/src/ui/panic.rs
@@ -1,6 +1,8 @@
 //! UI display for panic messages that occur outside of tests.
 
-use super::{font, load_ui_tiles, wait_for_vblank, Cursor, BG0CNT, BG1CNT, DISPCNT, TEXT_ENTRIES};
+use super::{
+    font, load_ui_tiles, wait_for_vblank, Cursor, BG0CNT, BG1CNT, DISPCNT, TEXT_ENTRIES, UI_ENTRIES,
+};
 use core::{fmt::Write, panic::PanicInfo};
 
 const DISPSTAT: *mut u16 = 0x0400_0004 as *mut u16;
@@ -29,6 +31,16 @@ pub(crate) fn display(info: &PanicInfo) -> ! {
     load_ui_tiles();
 
     wait_for_vblank();
+
+    // Clear previous text and highlights.
+    for y in 0..20 {
+        for x in 0..30 {
+            unsafe {
+                TEXT_ENTRIES.add(0x20 * y + x).write_volatile(0);
+                UI_ENTRIES.add(0x20 * y + x).write_volatile(0);
+            }
+        }
+    }
 
     let mut cursor = unsafe { Cursor::new(TEXT_ENTRIES) };
     write!(